ES/OS error causes without index data are handled properly (#14203) (#14263)
* ES/OS error causes without index data are handled properly

* Adding changelog entry
