+++ graylog2-server/src/test/java/org/graylog/storage/errors/CauseTest.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2020 Graylog, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the Server Side Public License, version 1,
+ * as published by MongoDB, Inc.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Server Side Public License for more details.
+ *
+ * You should have received a copy of the Server Side Public License
+ * along with this program. If not, see
+ * <http://www.mongodb.com/licensing/server-side-public-license>.
+ */
+package org.graylog.storage.errors;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertNull;
+
+class CauseTest {
+
+    @Test
+    void acceptsCauseWithoutIndexData() {
+        final Cause childishError = Cause.Builder.builder()
+                .reason("I won't give you data!")
+                .type("Childish error")
+                .build();
+
+        assertThat(childishError)
+                .satisfies(error -> assertNull(error.index()))
+                .satisfies(error -> assertNull(error.indexUuid()));
+    }
+}
